avoid error regarding folder not found

+++ nextcloud_update.sh
@@ -139,7 +139,7 @@ then
     rsync -Aaxz /var/log/ncvm/ "$VMLOGS"
     rm -Rf /var/log/ncvm/
     rm -f "$NCDATA"/*.log*
-elif [ -n "$(find /var/ncdata -maxdepth 1 -name "*.log")" ]
+elif [ -d /var/ncdata/] && [ -n "$(find /var/ncdata -maxdepth 1 -name "*.log")" ]
 then
     rsync -Aaxz /var/ncdata/*.log "$VMLOGS"
     rm -f /var/ncdata/*.log*
